refactor(richtext): move runs tests to RichStringRunsTest
